+++ storm-core/test/clj/org/apache/storm/drpc_test.clj
@@ -16,16 +16,15 @@
 (ns org.apache.storm.drpc-test
   (:use [clojure test])
   (:import [org.apache.storm.drpc ReturnResults DRPCSpout
-            LinearDRPCTopologyBuilder]
-           [org.apache.storm.utils ConfigUtils])
+            LinearDRPCTopologyBuilder])
   (:import [org.apache.storm.topology FailedException])
   (:import [org.apache.storm.coordination CoordinatedBolt$FinishedCallback])
   (:import [org.apache.storm LocalDRPC LocalCluster])
   (:import [org.apache.storm.tuple Fields])
-  (:import [org.apache.storm.utils.ConfigUtils])
+  (:import [org.apache.storm.utils ConfigUtils]
+           [org.apache.storm.utils.staticmocking ConfigUtilsInstaller])
   (:import [org.apache.storm.generated DRPCExecutionException])
   (:import [java.util.concurrent ConcurrentLinkedQueue])
-  (:import [org.apache.storm.testing.staticmocking MockedConfigUtils])
   (:use [org.apache.storm config testing clojure])
   (:use [org.apache.storm.daemon common drpc])
   (:use [conjure core]))
@@ -223,9 +222,10 @@
 (deftest test-dequeue-req-after-timeout
   (let [queue (ConcurrentLinkedQueue.)
         delay-seconds 2
-        conf {DRPC-REQUEST-TIMEOUT-SECS delay-seconds}]
-    (with-open [_ (proxy [MockedConfigUtils] []
-                      (readStormConfigImpl [] conf))]
+        conf {DRPC-REQUEST-TIMEOUT-SECS delay-seconds}
+        mock-cu (proxy [ConfigUtils] []
+                  (readStormConfigImpl [] conf))]
+    (with-open [_ (ConfigUtilsInstaller. mock-cu)]
       (stubbing [acquire-queue queue]
         (let [drpc-handler (service-handler conf)]
           (is (thrown? DRPCExecutionException
@@ -235,11 +235,12 @@
 (deftest test-drpc-timeout-cleanup 
   (let [queue (ConcurrentLinkedQueue.)
         delay-seconds 1
-        conf {DRPC-REQUEST-TIMEOUT-SECS delay-seconds}]
-    (with-open [_ (proxy [MockedConfigUtils] []
-                    (readStormConfigImpl [] conf))]
+        conf {DRPC-REQUEST-TIMEOUT-SECS delay-seconds}
+        mock-cu (proxy [ConfigUtils] []
+                  (readStormConfigImpl [] conf))]
+    (with-open [_ (ConfigUtilsInstaller. mock-cu)]
           (stubbing [acquire-queue queue
-               timeout-check-secs delay-seconds]
+                     timeout-check-secs delay-seconds]
               (let [drpc-handler (service-handler conf)]
                 (is (thrown? DRPCExecutionException 
                              (.execute drpc-handler "ArbitraryDRPCFunctionName" "no-args"))))))))